[REW] Started to rework the Window/Sample arch since we want interactions
As of right now, the sample is set to run for a given amount of frames.
For this demo, I would like to have the sample run until the windows is
closed by the user. To do this, I have to rework the way the sample and
window events are handled